Safari-609.1.11.1/LayoutTests/inspector/canvas/console-record-bitmaprenderer-expected.txt

Test that CanvasManager is able to record actions made to BitmapRenderer canvas contexts.


== Running test suite: Canvas.recordingBitmapRenderer
-- Running test case: Canvas.recordingBitmapRenderer.Console
PASS: The recording should have the name "TEST".
PASS: The recording should have one frame.
PASS: The first frame should have one action.
